Reorder parsing steps to fix an issue with brackets and delimiters
With a whitespace-delimited definition, the function s:BracketMatch
would remove any whitespace before checking for brackets, and that would
mean that an argument that starts with a quote or a bracket, would be
combined with the previous one.

One option would be to remove whitespace-stripping from the
s:BracketMatch function, but, really, I don't see a reason we wouldn't
check for a delimiter *before* checking for a bracket match, other than
because the matching itself needs to be done before the big if-clause
(which is also quite annoying, but it is what it is).