Rust: separate handling of function definitions and invocations
Function definitions have things like lifetimes (unbalanced single
quotes), but won't have comparisons. So, `<` and `> are brackets, and
`'` is not a bracket.

Function invocations don't have lifetimes (I think), so `'` is a
bracket, but `<` and `>` are not. This is not, strictly speaking
correct, because you can use turbofish as `Foo::<A, B>`. Will keep
thinking about this one, but I feel like it should be super rare.