Fix cursor positioning bug
After swapping the items, we move to a known item to re-parse them. The
goal is to figure out their new structure to know where to put the cursor.

However, when moving to the right, we were not necessarily putting the
cursor in the right *initial* place to reselect -- with a nested list,
the re-selection could result in parsing the inner one.

Example case:

  Result<Box<Error>, ()>

Solution is simply to always take the coordinate of the very first list
item. Whatever reshuffling we do, the original list of items' first
element will be the start of all the list items.